+++ android/app/src/main/java/com/zzang/chongdae/data/mapper/OfferingDetailResponseMapper.kt
@@ -1,10 +1,10 @@
 package com.zzang.chongdae.data.mapper
 
 import com.zzang.chongdae.data.remote.dto.response.OfferingDetailResponse
-import com.zzang.chongdae.domain.model.ArticleDetail
+import com.zzang.chongdae.domain.model.OfferingDetail
 
 fun OfferingDetailResponse.toDomain() =
-    ArticleDetail(
+    OfferingDetail(
         id = this.id,
         title = this.title,
         nickname = this.nickname,
