+++ android/app/src/main/java/com/zzang/chongdae/data/remote/dto/response/ParticipationResponse.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class ParticipationResponse(
     @SerialName("currentCount") val currentCount: Int,
-    @SerialName("status") val status: OfferingCondition,
+    @SerialName("status") val status: RemoteOfferingCondition,
 )
