+++ android/app/src/main/java/com/zzang/chongdae/data/remote/dto/response/RemoteOfferingCondition.kt
@@ -2,7 +2,7 @@ package com.zzang.chongdae.data.remote.dto.response
 
 import kotlinx.serialization.SerialName
 
-enum class OfferingCondition {
+enum class RemoteOfferingCondition {
     @SerialName("FULL")
     FULL,
