+++ android/app/src/main/java/com/zzang/chongdae/data/repository/remote/GroupPurchaseRepositoryImpl.kt
@@ -4,7 +4,7 @@ import com.zzang.chongdae.data.mapper.toDomain
 import com.zzang.chongdae.data.remote.dto.request.ParticipationRequest
 import com.zzang.chongdae.data.remote.source.GroupPurchaseDataSource
 import com.zzang.chongdae.domain.model.Article
-import com.zzang.chongdae.domain.model.ArticleDetail
+import com.zzang.chongdae.domain.model.OfferingDetail
 import com.zzang.chongdae.domain.model.Participation
 import com.zzang.chongdae.domain.repository.GroupPurchaseRepository
 
@@ -17,7 +17,7 @@ class GroupPurchaseRepositoryImpl(
         }
     }
 
-    override suspend fun fetchGroupPurchaseDetail(id: Long): Result<ArticleDetail> {
+    override suspend fun fetchGroupPurchaseDetail(id: Long): Result<OfferingDetail> {
         return groupPurchaseDataSource.fetchGroupPurchaseDetail(id = id).mapCatching {
             it.toDomain()
         }
