+++ android/app/src/main/java/com/zzang/chongdae/domain/model/OfferingCondition.kt
@@ -1,13 +1,13 @@
 package com.zzang.chongdae.domain.model
 
-enum class ArticleStatus {
+enum class OfferingCondition {
     FULL,
     TIME_OUT,
     CONFIRMED,
     AVAILABLE, ;
 
     companion object {
-        fun ArticleStatus.isAvailable(): Boolean {
+        fun OfferingCondition.isAvailable(): Boolean {
             return this == AVAILABLE
         }
     }
